net: lib: azure_iot_hub: Fix user_name generation
Currently, the `user_name` is generated in runtime only when DPS is used
and in all other cases it is a build-time concatenation of Kconfig config 
parameters. However this creates an issue when using a device id that is not 
set in Kconfig but is set in runtime by the app. This patch adds runtime 
generation for the user_name for this case.

Signed-off-by: Alex Tsamakos <alex@actinius.com>