lib: lte_lc: Parse only available info in CEREG notifications
Currently the link controller will log errors when trying to parse
CEREG notifications with level less than 4, as cell and PSM may
not be available.

This patch adds checks before attempting to parse CEREG notifications
to ensure that the information to be parsed is actually available.
In addition, PSM update events will now only be emitted if registered
with a network.

Fixes CIA-201

Signed-off-by: Jan Tore Guggedal <jantore.guggedal@nordicsemi.no>