Bluetooth: Mesh: Check of transition data/success allready deleted scenes.
Introduces a check of the transition time and delay, so the input values
can be checked and call will be rectected if they are outside bounderis.

The delete command of an allready deleted scene will now give success
instead of not found. This is according to the standard.

Added api functions for getting current and target scene, because
the scene state is callulated between prev, next and transition time.

Signed-off-by: Ingar Kulbrandstad <ingar.kulbrandstad@nordicsemi.no>