partition_manager: Fail/warn when for span/image name conflicts
Compiling a span as a partition image produces images with lengths
that are span-sized instead of sized appropriately for their image.
Linking with too long of lengths can lead to unintended image
overwriting. This can be avoided by naming container partitions
differently than the child image name.
For products already using this naming in pm_static.yml files,
print a warning about changing the names instead. A warning allows
awareness in developement while not blocking FW updates for these
products in the field.

Ref: NCSDK-4199, NCSDK-7982

Signed-off-by: Stephen Stauts <stephen.stauts@nordicsemi.no>
Signed-off-by: Francesco Servidio <francesco.servidio@nordicsemi.no>