dfu_target: fix bug with loading progress
Prior to this commit, the page to which the last
written offset belongs to would be erased.
This is incorrect since it has already been
erased.

Fix this by setting the 'last erased page'
variable in the stream context when
loading the bytes written

Also increase the name key buffer to be more
future proof.

Ref: NCSDK-7924
Signed-off-by: Håkon Øye Amundsen <haakon.amundsen@nordicsemi.no>