manifest: nrf_modem: version 1.0.0
With this version, the library has the possibility to configure the
shared memory layout. This patch adds support to do so via the glue.

The partition manager picks up the shared memory region sizes, and
reserves the necessary amount of RAM during linking, leaving the rest
available to the application.

In addition, the updated glue also implements the new functions:
- `nrf_modem_os_alloc`
- `nrf_modem_os_free`
- `nrf_modem_os_shm_tx_alloc`
- `nrf_modem_os_shm_tx_free`

using Zephyr's heap implementation.

The former two operate on a heap in the application RAM.
The latter two operate on a heap in the library's shmem TX region.
Both heaps are separate from Zephyr's system heap.

A few Kconfig options have been added to facilitate debugging.

Finally, this patch updates the offloading layer to support
newly introduced POSIX `getaddrinfo` error codes.

Signed-off-by: Emanuele Di Santo <emdi@nordicsemi.no>