nrf_desktop: Use CAF instead of dedicated implementation.
Make use of CAF in nrf_desktop application.
Buttons module is used for now.

Two events implementation are used in the application:
 - module_state_event
 - power_event

* Proper dependencies are introduced to modules that rely
on CAF from now.
* Application configuration files are updated to select CAF
options.
* Application CMakeLists.txt is slightly changed. Some of configuration
files required for CAF residue in ./configuration` directory.
CAF is build as a part of Zephyr and those configuration files must
be visible for cmake.

Signed-off-by: Emil Obalski <emil.obalski@nordicsemi.no>