applications: asset_tracker_v2: Remove broker hostname default string
Remove the default broker hostname string from the prj.conf. This will
cause the application to not build due to an assert statement in
AWS IoT. This is to give the user a strong indication that this config
must be set to the appropriate hostname, which is also reflected in the
documentation.

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>