applications: asset_tracker_v2: Align and add shutdown states to modules
Add shutdown states to modules. This prevents modules from handling
new events and queued events after an error has occurred.

This fixed the issue where an error would cause undefined behavior
across modules depending on the module shutdown order. This could give
a false impression of where the application failed in the first place.
The log would also get bloated with a lot of unnecessary error logs not
necessarily related to the original point of failure.

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>