mcuboot: mcuboot: add option for signature key file
If MCUBoot is not build from source, there is currently no way to
inform the build system about what key file should be used for
signing the application hex file. As a result of this, it is
currently not possible to build applications which enable MCUBoot
without building MCUBoot from source, as opposed to skipping
its build (if it is already programmed to the flash memory of
the device) or merging the application hex file with the MCUBoot
hex file.

This commit fixes this issue by adding a duplicate of the
BOOT_SIGNATURE_KEY_FILE kconfig option which is visible for the
application.

Ref: NCSDK-6859
Signed-off-by: Håkon Øye Amundsen <haakon.amundsen@nordicsemi.no>