samples: cloud_client: add missing configurations
By default, the cloud client sample uses the nRF cloud backend.
The nRF cloud uses the settings API to store an internal variable.
Without these configurations settings do not work properly and an error
is printed in the cloud backend when the library attempts to store the
variable to flash.

Closes [NCSDK-8311]

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>