sample: secure_services: read spm fw info instead of app
Passing address to non-secure memory is not allowed in
spm_firmware_info, also it is not a good test since this
memory is already readable by the app.

Instead, try to read the fw_info from the SPM address.
This is supported by spm_firmware_info, and verifies
that you can access data through the NSC which is not
available otherwise.

Ref: NCSDK-8348
Signed-off-by: Håkon Øye Amundsen <haakon.amundsen@nordicsemi.no>