applications: asset_tracker_v2: Check nrf_modem_lib_init return code
On boot the return code from nrf_modem_lib_init() should be checked
to determine if the modem requires a reboot.
A reboot is needed in the event of an ongoing modem firmware
update or an error happened during library initialization.

This patch prevents that the application proceeds and attempts
to send AT commands when the modem library is uninitialized,
resulting in a series of error messages and eventually a reboot.

Addresses CIA-239

Signed-off-by: Jan Tore Guggedal <jantore.guggedal@nordicsemi.no>