applications: asset_tracker_v2: Only send data values that are new.
Dynamic modem data is by default included in all packets published by
the device. However, a large portion of this type of data does not
change in between publications. This patch adds options and
functionality in the data and modem module that filters out parameters
if they have not changed since the last publication/sampling.

Initially, this involves dynamic modem data and the
device configuration. Data types such as battery,
accelerometer readings, and GPS coordinates are not affected by this PR
because they should be filtered based on ranges rather than their
previous value. This is a future improvement.

This PR includes:
        - Retaining of previous data values for dynamic modem data
          and device configuration.
	- Implement 'fresh' flags for parameters that are considered
	  updated and should be included in the encoding process.
        - Segmenting code into own static functions that handles the
          data in question. This is primarily to abstract away code from
          the event-handlers to make the code flow easier to read.
        - Add options to publish all data if desired. This is to provide
	  an option for users that do not want this new functionality.

Closes CIA-254

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>