cmake: add s1 variant of dev_handles.c
This fixes a bug where you would get orphan sections when
linking the S1 candidate.

All source code which is generated based on the first round
of linking needs to have the same handling in s1.cmake as
it does in zephyr/CMakeLists.txt.

This commit adds handling of generating 'dev_handles.c'
which was added in the latest upmerge.

Ref: NCSDK-9055
Signed-off-by: Håkon Øye Amundsen <haakon.amundsen@nordicsemi.no>