lib: lte_lc: Request eDRX and PTW separately for LTE-M and NB-IoT
It's now possible to have both LTE-M and NB-IoT enabled at the same
time, and let the modem choose which one to use based on user
preference configuration. That makes it hard to know for which
technology eDRX should be requested.
This patch introduces API changes requesting eDRX and PTW to let
the parameters be set individually for LTE-M and NB-IoT. In addition,
lte_lc_edrx_req() is updated to send the individual eDRX and PTW
values to the modem.

Signed-off-by: Jan Tore Guggedal <jantore.guggedal@nordicsemi.no>