Bluetooth: Mesh: Don't store/recall scene for extended models
This commit fixes bug where the extended models mess up Scene Recall
behaviour of the higher layer model. For example, both Generic Level and
Generic OnOff states are bound to Light Lightness state. When a scene is
recalled, both generic servers recalls Light Lightness state.

This change moves responsibility of storing/recalling states to the
higher layer models. The extension API helps to filter out extended
models when storing/recalling scene.

Responsibility of restoring model's last state at power-up is also
moved to the higher layer model. If a model is extended, it shall not
restore it's last state at start. This is also done through the
extension API.

Responsibility of invalidating a scene is moved to each model that has a
state marked as "Stored with Scene".

MeshMDLv1.0.1, section 5.1.3.1.1 also states the following:
If a model is extending another model, the extending model shall
determine the Stored with Scene behavior of that model.

Signed-off-by: Pavel Vasilyev <pavel.vasilyev@nordicsemi.no>