samples: bluetooth: periperal_hids_mouse: Use k_work for mouse events
Converts the current hids_work delayable work to a k_work item, as it's
not actually using the delay. This also means that the entry won't be
cancelled on disconnect, and remaining mouse events will be processed
after the connection is lost. This functionality is the same as it was,
but the intent to stop processing is no longer there.

Signed-off-by: Trond Einar Snekvik <Trond.Einar.Snekvik@nordicsemi.no>