net: lib: download_client: fix parsing header past end of buffer
http_header_parse() searches for the the HTTP header separator string
"\r\n\r\n" in client->buf, but doesn't limit the search to only the
data received in the current range response.
This change checks that if the first occurence of the separator string
is found past the end of the received data, we continue waiting for
the full HTTP header.

Signed-off-by: Mike Voytovich <mike@rokkresearch.com>