scripts: pygit2_helpers: support devicetree area
Add any commits whose shortlogs start with 'dtlib', 'edtlib', or
'devicetree' to the logical DT area of Zephyr development, and adjust
'Device tree' to 'Devicetree' in the area name.

Add 'Devicetree' to the allowlist of areas we expose to
get_pulls_in_range.py and adjust a linter nit while we're here.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>