partition manager: remove re-definition of BUILD_WITH_TFM
Although not an error, it seems redundant to have
a re-definition of BUILD_WITH_TFM option in the
Partition Manager's Kconfig, as it seems we only
need to have PM_SINGLE_IMAGE selected. Instead,
we set the PM_SINGLE_IMAGE as default y, when we
build with TF-M and make the symbol prompt-less.
We also add a relevant note in the help text of
the symbol, for clarity.

Signed-off-by: Ioannis Glaropoulos <Ioannis.Glaropoulos@nordicsemi.no>